Migrate to one component
